added BaseMap class with write/read methods
